chore: add author real info
