+++ tff-blog-api/Configuration/JsonSettings.cs
@@ -1,0 +1,14 @@
+﻿using System.Text.Json;
+
+namespace Tff.Blog.Api.Configuration
+{
+    public class JsonSettings
+    {
+        public static readonly JsonSerializerOptions Options = new()
+        {
+            PropertyNamingPolicy = JsonNamingPolicy.CamelCase,
+            PropertyNameCaseInsensitive = true,
+            WriteIndented = true,
+        };
+    }
+}
