+++ tff-blog-shared/Models/AuthorModel.cs
@@ -1,0 +1,9 @@
+﻿namespace Tff.Blog.Shared.Models
+{
+    public class AuthorModel : MarkdownModel
+    {
+        public string Photo { get; set; }
+
+        public string Name { get; set; }
+    }
+}
