+++ tff-blog-shared/Models/MarkdownModel.cs
@@ -1,0 +1,10 @@
+﻿namespace Tff.Blog.Shared.Models
+{
+    public abstract class MarkdownModel
+    {
+        /// <summary>
+        /// Text in markdown body
+        /// </summary>
+        public string Text { get; set; }
+    }
+}
