+++ tff-blog-shared/tff-blog-shared.csproj
@@ -8,4 +8,8 @@
     <AssemblyName>Tff.Blog.Shared</AssemblyName>
   </PropertyGroup>
 
+  <ItemGroup>
+    <PackageReference Include="YamlDotNet" Version="12.0.0" />
+  </ItemGroup>
+
 </Project>
