--- tff-blog/ViewModels/AuthorViewModel.cs
@@ -1,11 +1,0 @@
-﻿namespace Tff.Blog.ViewModels
-{
-    public class AuthorViewModel
-    {
-        public string PhotoUrl { get; set; }
-
-        public string Name { get; set; }
-
-        public string Description { get; set; }
-    }
-}
