--- tff-blog/wwwroot/data/author.json
@@ -1,5 +1,0 @@
-{
-  "photoUrl": "images/photo.jpg",
-  "name": "Tiago Fonseca",
-  "description": "An awesome developer! Lorem ipsum dolor sit amet, consectetur adipiscing elit. Quisque vulputate justo suscipit ex tempor ullamcorper. In ac scelerisque nunc, at lacinia tortor. Nam pharetra sollicitudin est interdum varius. In eros enim, sodales tempus ultrices nec, commodo ac lacus. Ut porttitor libero et leo faucibus, sed ornare dui condimentum. Ut quis pharetra est. Phasellus arcu odio, facilisis ac facilisis quis, facilisis vel ipsum."
-}
