+++ tff-blog/wwwroot/pages/author.md
@@ -1,0 +1,6 @@
+﻿---
+photo: images/photo.jpg
+name: Tiago Fonseca
+---
+Developer 👨‍💻 | Learning Addict 📚 
+Passionate about solving business problems using clean solutions
